requirements.txt
discord.py==1.2.3
youtube-dl==2019.7.16
